+++ play.py
@@ -1,5 +1,4 @@
 from game.cli import Game
 
 game = Game()
-
 game.cli()